refactor: rename the strategy files to match classes
