+++ packages/core/src/index.ts
@@ -5,8 +5,8 @@ export * from "./errors";
 export { IgnitionModuleSerializer } from "./ignition-module-serializer";
 export { formatSolidityParameter } from "./internal/formatters";
 export { status } from "./status";
-export { BasicStrategy } from "./strategies/basic-execution-strategy";
-export { Create2Strategy } from "./strategies/create2-execution-strategy";
+export { BasicStrategy } from "./strategies/basic-strategy";
+export { Create2Strategy } from "./strategies/create2-strategy";
 export * from "./type-guards";
 export * from "./types/artifact";
 export * from "./types/deploy";
