+++ packages/core/test/execution/future-processor/utils.ts
@@ -18,7 +18,7 @@ import {
 import { getDefaultSender } from "../../../src/internal/execution/utils/get-default-sender";
 import { MemoryJournal } from "../../../src/internal/journal/memory-journal";
 import { assertIgnitionInvariant } from "../../../src/internal/utils/assertions";
-import { BasicStrategy } from "../../../src/strategies/basic-execution-strategy";
+import { BasicStrategy } from "../../../src/strategies/basic-strategy";
 import {
   exampleAccounts,
   setupMockArtifactResolver,
